tools/strat_visualizer/requirements.txt
pygame~=2.6.1
paho-mqtt~=2.1.0
numpy~=2.2.3
shapely~=2.0.7
